Add more samples data

+++ BeerProject/src/main/java/wawer/kamil/beerproject/domain/Beer.java
@@ -26,7 +26,7 @@ public class Beer implements Serializable {
 
     private String style;
 
-    private Integer extract;
+    private Double extract;
 
     private Double alcohol;
